scs_server: add commit titles and messages to blame results
Summary:
Avoid clients needing to make a secondary look-up for the commit title or message, which
we already have, by returning it in the blame response if the clients request it.

To do this, we add `format_options` to the blame request parameters, which allows customization
of the returned response.  The default value if not specified corresponds to the old
behaviour for backwards compatibility.

Reviewed By: mitrandir77

Differential Revision: D25057231

fbshipit-source-id: 84faad7b2db1f684beb94d9fa799c8cef1a89ce8