admin: add skeleton manifests to admin command
Summary:
Add skeleton manifests to `derived-data verify-manifests` and add a
`skeleton-manifests` subcommand that allows listing and traversing of skeleton
manifests.

Since skeleton manfiests don't contain the content id or file type, they are
valid simply if they contain the same set of files as the manifest we are
comparing against.

To allow testing of these items while backfilling is still in progress, add an
`--if-derived` option to these commands that work even if derivation is
disabled, provided the manifest has already been derived.

Reviewed By: ikostia

Differential Revision: D25064785

fbshipit-source-id: a6f923bfc53262a5b2118917f8fdd3e99407e036