fsnodes: don't rederive parents fsnodes during batch derivation
Summary:
On the second and subsequent linear stack in a batch, the parent fsnode may
have been derived in the previous iteration of the loop.  Since we haven't
completed this batch yet, the mappings have not been stored, and so the
attempt to derive the parent will result in the parent being derived again,
repeating the work of this batch sequentially.

Apply the optimization used in skeleton manifests and fetch the parent fsnodes
out of the result we are accumulating.

Reviewed By: StanislavGlebik

Differential Revision: D25218194

fbshipit-source-id: 5cc49204b53984f8aa73542f1a794a6251eb2b2e