mononoke/unbundle: stop collecting ContentBlobMeta given we never use it
Summary:
Like it says in the title. We've had this comment about potentially using this
for a couple of years now. It seems a bit unlikely at this point that we're
going to use this, but currently it makes the code that provides uploading hg
entries more complex than it needs to be, so let's just get rid of this.

Reviewed By: lukaspiatkowski, StanislavGlebik

Differential Revision: D25219728

fbshipit-source-id: 61d254edef16d24a4c29f96f983f894863b5232a