mononoke/bonsai_hg_mapping: update to futures 0.3 get_or_fill
Summary:
Like it says in the title, this updates bonsai_hg_mapping to a futures 0.3
implementation of get_or_fill.

The upshot is that this requires less cloning (in fact, no cloning at all if
the rest of the code was 0.3 here), and in this particular instance it also
lets us completely get rid of the `from_bonsai` flag we were threading through
the whole method and checking many times.

Reviewed By: StanislavGlebik

Differential Revision: D25396199

fbshipit-source-id: f8126c96aad8d982c3deb535530484bec841929f